refactor(kubernetes): Minor refactor of timestamp function
Instead of looking twice in the map, just check if the result is
null in getCreationTimestamp.  (Since this is a mutable map, the
presence of the key doesn't guarantee the value is null anyway, so
let's jut get the value.) I didn't use Optional here because this
is used in a comparator that might be called a lot for large sorts,
but maybe that's a micro-optimization.

The result is non-null so we can use .isEmpty() in getFormattedTimestamp.

Finally let's rename getFormattedTimestamp to getCreationTimestampEpochMillis,
as the result is not formatted.