refactor(kubernetes): Fix JdkObsolete warnings
[JdkObsolete] Date has a bad API that leads to bugs; prefer
java.time.Instant or LocalDate.

In KubernetesManifest, we can parse the String into an Instant
instead of a Date. (There's also no point separately handling
an empty string; anything invalid will cause a ParseException
so just let that happen.)

In KubernetesModelUtil, the only place that was calling this
function already had a org.joda.time.DateTime object and was
serializing it before calling this function; we can just directly
get the milliseconds from the object and delete translateTime.