refactor(kubernetes): Delete tautological tests
After the last refactor, some of the tests on the artifact converters
are just testing Artifact itself; there's no need for us to ensure
that creating an artifact with a reference returns that reference.

We should replace these with better tests, but I'm going to wait until
a later commit when logic is further consolidated.