refactor(kubernetes): Return OptionalInt from getVersion
The version parsing logic should return an OptionalInt instead of
a String; this doesn't help us yet as we're just going to convert
it to a String in getArtifact, but this will enable us in a later
commit to directly access the int.

Also, unrelatedly, move a few variable declarations closer to where
they are used in KubernetesDeployManifestOperation.