refactor(kubernetes): Split versioning and artifact creation
Create a new ResourceVersioner class that is effectively
the logic that was in KubernetesVersionedArtifactConverter (in
fact the diff should even show it as a move). We'll make this a
component so we don't need to always pass in the ArtifactProvider
and can simplify the API. This class is responsible for taking a
KubernetesManifest and telling you what version should be appended
to its name.

Have KubernetesArtifactConverter require a version as input; this gets
it out of the business of figuring out what version should be applied
and with the core competency of just creating an artifact. It now
just has the single toArtifact function, which we'll make static and
remove the (unecessary) implementation classes.

The remaining commits are just changing from passing around an
ArtifactProvider to passing around a ResourceVersioner.