refactor(kubernetes): make logical and cluster relationships agnostic to Kubernetes kind
We map Kubernetes kinds to Spinnaker kinds in order to surface them in the Application UI as Clusters, Load Balancers, and Firewalls. Looking ahead to when CRDs may be mappable to Spinnaker kinds as well, let's not keep a fragile list of Kubernetes kinds that have these Spinnaker-kind-specific relationships, and instead just check by Spinnaker kind.

(This commit is not necessary to solve the problem this PR set out to solve, so if y'all think this actually makes it less clear or it's too cumbersome to pass around the SpinnakerKubernetesKindMap, happy to remove it from this PR!)