fix(servergroups): add `account` property when expand=true
When calling `application/{}/serverGroups?expand=false` clouddriver returns an account field as part of the servergroupviewmodel.
However, when calling the same with `expand=true` the `account` field is missing and isntead we have an `accountName` field representing same information

This if difficult to deal with for API users. Adding, `account` field for consistency