chore(titus): Log duplicate Titus server group names by job ID
We have seen an issue wherein competing in-flight requests to Titus result in multiple jobs with the same server group name.  This is due to a race condition in the Titus server group name validation handling, and something that will be addressed in Titus.  In any case, we can log and alert on this error.