fix(entitytags): include required module
The removal of the AWS integration with Entity Tags as part of #4923 mean that the required modules for core Entity Tags functions was not included in the `clouddriver` build.

This change changes the dependency on `clouddriver-elasticsearch` in `clouddriver-web` from test only to implementation. This ensure that the ES module is included in the build.

Fixes spinnaker/spinnaker#6331