feat(kubernetes): Breaking, Update API versions (#5275)
* feat(kubernetes): Breaking, Update API versions

This change makes Spinnaker incompatible with Kubernetes prior to 1.14.

The reason for this change is to enable the now General Available
Ingress API. This was added in Kubernetes 1.19. To make this change the
client library had to be updated. Since 1.16 removed a bunch of APIs
that change now affects us, and a few of the old APIs got removed.

As far as I can tell the oldest supported version of Kubernetes is now
1.14. It was released in 2019, almost two years ago to date. It's been
End of Life since December 2019.

BREAKING CHANGE: This breaks Kubernetes prior to 1.14.

* feat(kubernetes): Implement requested changes

This change implements the requested changes from the PR review