feat(kubernetes): Opt-out of artifact binding
In "Deploy (Manifest)" stage there's a new configuration parameter called "enableArtifactBinding" defaulting to "true", which controls whether binding of required or optional artifacts should be done for the manifests being deployed. Default behavior is backwards compatible with existing pipelines (automatic binding enabled), and the parameter allows for users to opt-out of automatic binding. Automatic binding of versioned objects (ConfigMap, Secrets) that are deployed as part of a single multidoc yaml document is still done regardless of the config parameter.