fix(cache/sql): improve behavior when attempting to cache an item with a duplicate id
Previously, given items to cache in the same order, the contents of the cache would cycle
through the items.  With this change, the contents of the cache is predictable -- the
first item.

It's still not ideal for legitimately different items to have the same id (e.g. two AMIs
with the same name in the same account), but at least now the cache stays constant instead
of changing, say, each time an ImageCachingAgent runs.