fix(entity tags): change RetrofitError and NotFoundException handling to SpinnakerHttpException with response code 404
Depends on https://github.com/spinnaker/kork/pull/911.

This is similar to https://github.com/spinnaker/clouddriver/pull/4631, but does actually
change behavior.  Instead of deleting from ElasticSearch on any RetrofitError, now the
code only deletes from ElasticSearch when the entity tags aren't found in front50.  This
seems to be the intention of the code, but it is a change.