fix(provider/azure): Fix CreateAzureServerGroupWithAzureLoadBalancerAtomicOperation and DestroyAzureServerGroupAtomicOperation after migrating to latest Azure SDK
After the migration of the SDK to com.azure.resourcemanager:azure-resourcemanager:2.19.0 CreateAzureServerGroupWithAzureLoadBalancerAtomicOperation
started to fail because the enableServerGroupWithLoadBalancer operation is not able to update the load-balancing rule

The reason is that the old SDK, defines only one backendAddressPool property in LoadBalancingRuleInner. But with the new SDK, two properties are defined backendAddressPool & backendAddressPools.  The atomic operation updates the backendAddressPool to point to the new server group, while the list contains the default backend pool.  Because of this, Azure API is throwing LoadBalancingRuleBackendAdressPoolAndBackendAddressPoolsCannotBeSetAtTheSameTimeWithDifferentValue exception. The fix is to set both properties to use the same value.

Closes https://github.com/spinnaker/spinnaker/issues/6736