perf(cache): Optimise heap usage in SqlCache
This pull request addresses a performance issue discovered in the Clouddriver, specifically within the `SqlCache.getHash method`.

I used async-profiler to build an Allocation profile and a CPU profile. From the Allocation profile, it was found that the `SqlCache.getHash` method was responsible for 80% of the memory allocation, mostly due to the heavy use of `String.format` within this method. Furthermore, the CPU profile indicated that 34% of CPU usage was also attributed to this method.

So I'm making a conclusion that `SqlCache.getHash` is generating a significant amount of garbage, contributing to high heap and CPU usage. This could potentially be a root cause for Out Of Memory errors if the Garbage Collector is unable to clean up the generated garbage efficiently. It also may add to high CPU usage.
Perhaps, relevant issues: [#6835](https://github.com/spinnaker/spinnaker/issues/6835), [#6797](https://github.com/spinnaker/spinnaker/issues/6797), and [#6810](https://github.com/spinnaker/spinnaker/issues/6810).

The proposed optimisation in this pull request aims to reduce the amount of garbage generated by this method by pre-generating string representations for all possible byte values and using `StringBuilder` with a predefined size.

This optimisation has been deployed in our production environment for a month and has shown promising results, reducing resource consumption by 2-5 times. I attached the screenshots of the graphs to this pull request for illustration.