feat(kubernetes): Support multiple kubectl versions
Currently it is hard to support clusters that runs different Kubernetes versions. Because kubectl executables can be defined per cluster, we can overcome this limitation by installing multiple kubectl versions, and then define which binary to use in the cluster config.
The kubectl versions are stored without patch version, so that we can upgrade to newer versions without having to change the configuration. We also symlink `kubectl` to the current default version to not break existing configuration.

To opt in to a newer version, define it like this in the cluster config: `kubectlExecutable: /usr/local/bin/kubectl-1.28`
E.g.:
```
kubernetes:
  enabled: true
  primaryAccount: my-cluster
  accounts:
    - name: my-cluster
      kubeconfigFile: configserver:k8s/my-cluster.yml
      cacheThreads: 1
      cachingPolicies: []
      checkPermissionsOnStartup: true
      customResources: []
      dockerRegistries: []
      kinds: []
      liveManifestCalls: false
      namespaces: []
      oAuthScopes: []
      omitKinds: []
      omitNamespaces: []
      onlySpinnakerManaged: false
      providerVersion: V2
      kubectlExecutable: /usr/local/bin/kubectl-1.28
```