Add test for Kubernetes job pod deleted
When running a Kubernetes job stage it is possible for the pod to be
deleted between the job succeeding and clouddriver running the
`collectJob` function.  When this happens clouddriver hits an
`IndexOutOfBoundsException` error as it attempts to get the name
of the deleted pod.

Even if the pod is deleted we want the stage to succeed as the job
itself succeeded and that should be our measure of success.

Co-authored-by: dpachev <dimitar.pachev@skyscanner.net>